seat: move pointing device and switch device to separate classes
