+++ src/core/seat/seat.hpp
@@ -29,27 +29,10 @@ class wf_input_device_internal : public wf::input_device_t
 {
     public:
     wf_input_device_internal(wlr_input_device* dev);
+    virtual ~wf_input_device_internal() = default;
 
-    wf::wl_listener_wrapper on_switch, on_destroy;
-
-    void handle_switched(wlr_event_switch_toggle *ev);
-    void update_options();
-
-    static struct config_t
-    {
-        wf_option mouse_cursor_speed;
-        wf_option mouse_scroll_speed;
-        wf_option touchpad_cursor_speed;
-        wf_option touchpad_scroll_speed;
-        wf_option touchpad_tap_enabled;
-        wf_option touchpad_click_method;
-        wf_option touchpad_scroll_method;
-        wf_option touchpad_dwt_enabled;
-        wf_option touchpad_dwmouse_enabled;
-        wf_option touchpad_natural_scroll_enabled;
-
-        void load(wayfire_config *config);
-    } config;
+    wf::wl_listener_wrapper on_destroy;
+    virtual void update_options() {};
 };
 
 /** Convert the given point to a surface-local point */
